Automatically switch to byte mode to enable Cesanta bootloader
The new esptool.py loads a special program in memory and then runs it.
This special program requires byte mode (it does not use the frames of
the esp rom) and so we now switch back to byte mode after seeing a
MEM_END of FLASH_END command.

Updated esptool to the most recent version from github.

Also improved the visual feedbacks on the leds.