Even more code reorganization
There does not seem to be a true benefit to moving shared files in a
library when we have to split it into multiple pieces and then have a
complex build config to make everything work.

Instead, I hope moving everything into src/ will make it easier to
understand the architecture of the source, and to understand the build
configuration (which is in platformio.ini).